requirements.txt
compressai==1.1.3
imageio==2.9.0
matplotlib==3.3.4
numpy==1.20.1
opencv-python==4.5.1.48
pathlib2==2.3.5
Pillow @ file:///tmp/build/80754af9/pillow_1609786786540/work
piq==0.5.5
pytorch-msssim==0.2.1
PyYAML==5.4.1
scikit-image==0.18.1
scipy==1.6.0
torch==1.10.0+cu113
torchac==0.9.3
torchaudio==0.10.0+cu113
torchsearchsorted @ file:///home/chajin/torchsearchsorted
torchvision==0.11.1+cu113
tqdm==4.56.1

